basic implementation of the meta tokenizer
