+++ gtars/tests/data/chroms.meta.bed
@@ -1,0 +1,24 @@
+chr1	0	248956422	1
+chr2	0	242193529	1
+chr3	0	198295559	2
+chr4	0	190214555	2
+chr5	0	181538259	3
+chr6	0	170805979	3
+chr7	0	159345973	4
+chr8	0	145138636	4
+chr9	0	138394717	5
+chr10	0	133797422	5
+chr11	0	135086622	6
+chr12	0	133275309	6
+chr13	0	114364328	6
+chr14	0	107043718	7
+chr15	0	101991189	7
+chr16	0	90338345	8
+chr17	0	83257441	8
+chr18	0	80373285	8
+chr19	0	58617616	8
+chr20	0	64444167	9
+chr21	0	46709983	9
+chr22	0	50818468	10
+chrX	0	156040895	11
+chrY	0	57227415	11
