+++ gtars/tests/data/tokenizer.meta.toml
@@ -1,0 +1,1 @@
+universes = ["peaks.meta.bed", "chroms.meta.bed"]
